+++ tests/unit/style/size_of.rs
@@ -12,8 +12,3 @@ size_of_test!(test_size_of_property_declaration, properties::PropertyDeclaration
 // This is huge, but we allocate it on the stack and then never move it,
 // we only pass `&mut SourcePropertyDeclaration` references around.
 size_of_test!(test_size_of_parsed_declaration, properties::SourcePropertyDeclaration, 576);
-
-#[test]
-fn size_of_specified_values() {
-    ::style::properties::test_size_of_specified_values();
-}
