+++ tests/unit/stylo/size_of.rs
@@ -47,8 +47,3 @@ size_of_test!(test_size_of_rule_node, RuleNode, 80);
 // This is huge, but we allocate it on the stack and then never move it,
 // we only pass `&mut SourcePropertyDeclaration` references around.
 size_of_test!(test_size_of_parsed_declaration, style::properties::SourcePropertyDeclaration, 704);
-
-#[test]
-fn size_of_specified_values() {
-    ::style::properties::test_size_of_specified_values();
-}
